MAP-1060: more work on QA coverage.
