+++ admin/updatedb/pom.xml
@@ -15,13 +15,6 @@
   <description>OTF RF2 Administrator Tool to Update the Database</description>
   <packaging>pom</packaging>
 
-  <properties>
-    <!-- This can be overridden to use "create" instead via -D....=create -->
-    <!-- NOTE: this setting alone, as a maven property filters down to the HBM2DDL 
-      tool and causes it to work, no need for environment or config file properties -->
-    <hibernate.hbm2ddl.auto>update</hibernate.hbm2ddl.auto>
-  </properties>
-
   <profiles>
 
     <!-- Activated by run.config setting -->
