+++ integration-tests/src/test/java/org/ihtsdo/otf/mapping/test/other/CompareMapRecordJpaTest.java
@@ -1,4 +1,4 @@
-package org.ihtsdo.otf.mapping.jpa.services;
+package org.ihtsdo.otf.mapping.test.other;
 
 import static org.junit.Assert.fail;
