+++ integration-tests/src/test/java/org/ihtsdo/otf/mapping/test/other/WorkflowPathStatesNonLegacyPathTest.java
@@ -1,4 +1,4 @@
-package org.ihtsdo.otf.mapping.jpa.services;
+package org.ihtsdo.otf.mapping.test.other;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
@@ -19,7 +19,10 @@ import org.ihtsdo.otf.mapping.helpers.WorkflowType;
 import org.ihtsdo.otf.mapping.jpa.MapProjectJpa;
 import org.ihtsdo.otf.mapping.jpa.MapRecordJpa;
 import org.ihtsdo.otf.mapping.jpa.MapUserJpa;
-import org.ihtsdo.otf.mapping.jpa.handlers.WorkflowFixErrorPathHandler;
+import org.ihtsdo.otf.mapping.jpa.handlers.WorkflowNonLegacyPathHandler;
+import org.ihtsdo.otf.mapping.jpa.services.ContentServiceJpa;
+import org.ihtsdo.otf.mapping.jpa.services.MappingServiceJpa;
+import org.ihtsdo.otf.mapping.jpa.services.WorkflowServiceJpa;
 import org.ihtsdo.otf.mapping.model.MapProject;
 import org.ihtsdo.otf.mapping.model.MapRecord;
 import org.ihtsdo.otf.mapping.model.MapUser;
@@ -34,12 +37,12 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 /**
- * Unit test for workflow path states for "fix error path".
+ * Unit test for workflow path states on "non legacy path".
  */
-public class WorkflowPathStatesFixErrorPathTest {
+public class WorkflowPathStatesNonLegacyPathTest {
 
   /**  The handler. */
-  private static WorkflowFixErrorPathHandler handler;
+  private static WorkflowNonLegacyPathHandler handler;
 
   /**  The content service. */
   private static ContentService contentService;
@@ -54,7 +57,7 @@ public class WorkflowPathStatesFixErrorPathTest {
   private static TrackingRecord trackingRecord;
 
   /**  The lead. */
-  private static MapUser specialist, lead;
+  private static MapUser specialist, specialist2, lead;
 
   /**  The map project. */
   private static MapProject mapProject;
@@ -78,7 +81,7 @@ public class WorkflowPathStatesFixErrorPathTest {
     workflowService = new WorkflowServiceJpa();
 
     // instantiate the workflow handler
-    handler = new WorkflowFixErrorPathHandler();
+    handler = new WorkflowNonLegacyPathHandler();
 
     // ensure database is clean
     for (Concept c : contentService.getConcepts().getIterable())
@@ -111,6 +114,13 @@ public class WorkflowPathStatesFixErrorPathTest {
     specialist.setUserName("spec");
     mappingService.addMapUser(specialist);
 
+    specialist2 = new MapUserJpa();
+    specialist2.setApplicationRole(MapUserRole.VIEWER);
+    specialist2.setEmail("none");
+    specialist2.setName("Specialist2");
+    specialist2.setUserName("spec2");
+    mappingService.addMapUser(specialist2);
+
     lead = new MapUserJpa();
     lead.setApplicationRole(MapUserRole.VIEWER);
     lead.setEmail("none");
@@ -135,7 +145,7 @@ public class WorkflowPathStatesFixErrorPathTest {
     mapProject.setPublic(true);
     mapProject.setRefSetId("refsetId");
     mapProject.setRuleBased(true);
-    mapProject.setWorkflowType(WorkflowType.REVIEW_PROJECT);
+    mapProject.setWorkflowType(WorkflowType.CONFLICT_PROJECT);
     mapProject.addMapSpecialist(specialist);
     mapProject.addMapLead(lead);
     mapProject.addScopeConcept("1");
@@ -151,8 +161,9 @@ public class WorkflowPathStatesFixErrorPathTest {
   @Test
   public void testLegalWorkflowCombinations() throws Exception {
 
-    Logger.getLogger(WorkflowPathStatesFixErrorPathTest.class)
-        .info("Testing all possible combinations against legal states...");
+    Logger.getLogger(WorkflowPathStatesNonLegacyPathTest.class).info(
+        "Testing all possible combinations against legal states ("
+            + handler.getWorkflowStatusCombinations().size() + "found) ...");
 
     // test empty state
     if (handler.isEmptyWorkflowAllowed()) {
@@ -245,7 +256,8 @@ public class WorkflowPathStatesFixErrorPathTest {
                 .equals(WorkflowStatus.PUBLISHED)) {
           // do nothing
         } else {
-          fail("Error computing tracking record for combination " + combination.toString());
+          fail("Error computing tracking record for combination "
+              + combination.toString());
         }
 
       }
@@ -262,7 +274,7 @@ public class WorkflowPathStatesFixErrorPathTest {
   @AfterClass
   public static void cleanup() throws Exception {
 
-    Logger.getLogger(WorkflowPathStatesFixErrorPathTest.class)
+    Logger.getLogger(WorkflowPathStatesNonLegacyPathTest.class)
         .info("Clean-up");
     resetRecords();
 
@@ -331,38 +343,38 @@ public class WorkflowPathStatesFixErrorPathTest {
     resetRecords();
 
     // sleep 0.5s to allow transaction to complete
-    Thread.sleep(500);
+    Thread.sleep(1000);
 
     Iterator<WorkflowStatus> statusIter =
         combination.getWorkflowStatusesAsList().iterator();
 
     // switch on size of combination
-    switch (combination.getWorkflowStatuses().size()) {
+    switch (combination.getWorkflowStatusesAsList().size()) {
 
-      // always invalid
+    // empty workflow
       case 1:
         mappingService
-        .addMapRecord(createRecord(specialist, statusIter.next()));
+            .addMapRecord(createRecord(specialist, statusIter.next()));
         break;
-     // valid: REVISION + Specialist editing/complete
       case 2:
         mappingService
-        .addMapRecord(createRecord(specialist, statusIter.next()));
-        mappingService
             .addMapRecord(createRecord(specialist, statusIter.next()));
+        mappingService
+            .addMapRecord(createRecord(specialist2, statusIter.next()));
         break;
-        // valid: REVISION + Specialist omplete + Lead editing/complete
       case 3:
         mappingService
-        .addMapRecord(createRecord(specialist, statusIter.next()));
-        mappingService
             .addMapRecord(createRecord(specialist, statusIter.next()));
+        mappingService
+            .addMapRecord(createRecord(specialist2, statusIter.next()));
         mappingService.addMapRecord(createRecord(lead, statusIter.next()));
         break;
       default:
         fail("Unexpected number of workflow statuses, combination = "
             + combination.toString());
     }
+
+    Thread.sleep(500);
 
     // compute workflow
     workflowService.computeWorkflow(mapProject);
